// __handoff_v21/J-Flow/electron/types.ts
export interface SyncInfoFile {
  syncVersion: number
  createdAt: string
  updatedAt: string
  appName: 'J-Flow'
  minSupportedAppVersion: string
}

export interface SyncDeviceInfoFile {
  syncVersion: number
  deviceId: string
  deviceName: string
  platform: string
  appVersion: string
  lastSeenAt: string
  lastSyncedAt: string | null
}